feat(componentbase): Advances in state manager, persistent data output and informative warnings
